+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Removed the Queue::sort method
 * Added Judy as the Database array.
 * save_signal_history rennamed to set_signal_history
-* Processor::search_signals renamed to Processor::find_signal
+* Processor::search_signals renamed to Processor::find_signal_database
 * State::get/set_state are now final.
 * signal\Standard renamed to SIG_Base
 * Signal renamed to SIG